ml-services/puzzle-solver/requirements.txt
flask==3.1.0
opencv-python-headless==4.9.0.80
numpy>=1.24.0
Pillow>=10.0.0
gunicorn>=21.2.0
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.15.0
requests>=2.31.0
